Переход на BigDecimal дял всех сумм денег - добавлен лог изменений бд - исправлены все поля связанные с деньгами - изменены методы под новый параметр

+++ src/main/java/ru/yandex/finance_tracker/dto/output/AccountInfoDto.java
@@ -5,6 +5,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.yandex.finance_tracker.model.Currency;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ public class AccountInfoDto {
     private Integer id;
     private String name;
     private Currency currency;
-    private Float balance;
+    private BigDecimal balance;
 }
