+++ src/main/java/ru/yandex/finance_tracker/dto/output/CategoryExpenseDto.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CategoryExpenseDto {
     private String category;
-    private Float totalExpense;
+    private BigDecimal totalExpense;
 }
