+++ src/main/java/ru/yandex/finance_tracker/dto/output/TransactionInfoDto.java
@@ -3,17 +3,18 @@ package ru.yandex.finance_tracker.dto.output;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import ru.yandex.finance_tracker.model.Type;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record TransactionInfoDto(
         Long transactionId,
         Long accountId,
         Type type,
-        Float amount,
+        BigDecimal amount,
         String category,
         @JsonFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
         String description,
-        Float accountBalance
+        BigDecimal accountBalance
 ) {
 }
